# openapi-spec.yaml
openapi: "3.0.3"
info:
  title: 韧性云 - 公共安全管理平台 API
  description: |
    韧性云公共安全管理平台后端接口文档。

    ## 认证
    除登录接口外，大部分接口需 Bearer Token（JWT），在请求头中携带：
    ```
    Authorization: Bearer <token>
    ```

    ## 统一响应格式
    ```json
    { "code": 0, "message": "ok", "data": <具体数据> }
    ```
    - `code: 0` 成功，`code: 400` 参数错误，`code: 401` 未登录
  version: "1.0.0"
  contact:
    name: 韧性云开发团队
servers:
  - url: http://localhost:3201
    description: 本地开发环境
  - url: /api
    description: 前端代理（Vite proxy → :3201）
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EnterpriseUpdate:
      type: object
      properties:
        name:
          type: string
        dimALevel1:
          type: string
        dimALevel2:
          type: string
        dimALevel3:
          type: string
        dimB:
          type: string
        dimCCode:
          type: string
        dimCName:
          type: string
        dimD:
          type: string
        region:
          type: string
        contactName:
          type: string
        contactPhone:
          type: string
        validFrom:
          type: string
        validTo:
          type: string
        address:
          type: string
        remark:
          type: string
        logo:
          type: string
        mapLng:
          type: number
        mapLat:
          type: number
        mapAddress:
          type: string

    # ---- 企业关联 ----
...
paths:
...
  /api/enterprise/{id}:
    get:
      tags:
        - 租户管理
      summary: 企业详情
      description: 获取企业完整信息（需运营管理或技术管理权限）
      operationId: getEnterpriseDetail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

    put:
      tags:
        - 租户管理
      summary: 更新企业
      operationId: updateEnterprise
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseUpdate'
      responses:
        "200":
          description: OK
...